+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 4.4.1 (unreleased)
 
 - Added `stem_exclusion` and `stemmer_override` options
+- Added `with_score` method to search results
 - Improved error message for `reload_synonyms` with non-OSS version of Elasticsearch
 - Improved output for reindex rake task